- 32 vs. 64 bit: ioctl command is u_long, not int

+++ sbin/ifconfig/ifconfig.c
@@ -6448,7 +6448,7 @@ clrlabel(const char *val, int d)
 void
 process_label_commands(void)
 {
-	int	cmd;
+	u_long	cmd;
 	const char *cmd_name;
 
 	switch (actions & (A_LABELSET|A_LABELCLR)) {
